"ads2_m8_a1_1: 6/6 testcases passed. Cannot run hidden testcases as Server is unreachablecheck_style score: 1/1 "

+++ ADS-2 Assignments/M8/1. Shortest Paths/Solution.java
@@ -41,7 +41,7 @@ public final class Solution {
             DijkstraUndirectedSP(ewgobj, source);
             if (dijkstra.hasPathTo(hashobj.get(srcDest[1]))) {
                 System.out.println(
-                    dijkstra.distTo(hashobj.get(srcDest[1])));
+                   (int) dijkstra.distTo(hashobj.get(srcDest[1])));
             }
             queries--;
         }
